backend/requirements.txt
# FastAPI and related dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database dependencies
sqlalchemy==2.0.23
alembic==1.12.1
pymysql==1.1.0
cryptography==41.0.7

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP client for external API calls
httpx==0.25.2
requests==2.31.0

# Date and time utilities
python-dateutil==2.8.2

# Machine Learning dependencies (lightweight)
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4

# Text-to-Speech (for server-side TTS if needed)
pyttsx3==2.90

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# CORS for frontend communication
fastapi-cors==0.0.6
